Example cleanup (#580)

+++ newton/tests/test_examples.py
@@ -78,6 +78,7 @@ def add_example_test(
     test_options: dict[str, Any] | None = None,
     test_options_cpu: dict[str, Any] | None = None,
     test_options_cuda: dict[str, Any] | None = None,
+    use_viewer: bool = False,
 ):
     """Registers a Newton example to run on ``devices`` as a TestCase."""
 
@@ -137,21 +138,32 @@ def add_example_test(
         # Append Warp commands
         command.extend(["-m", f"newton.examples.{name}", "--device", str(device)])
 
-        stage_path = (
-            options.pop(
-                "stage_path",
-                os.path.join(os.path.dirname(__file__), f"outputs/{name}_{sanitize_identifier(device)}.usd"),
+        if not use_viewer:
+            stage_path = (
+                options.pop(
+                    "stage_path",
+                    os.path.join(os.path.dirname(__file__), f"outputs/{name}_{sanitize_identifier(device)}.usd"),
+                )
+                if USD_AVAILABLE
+                else "None"
             )
-            if USD_AVAILABLE
-            else "None"
-        )
 
-        if stage_path:
-            command.extend(["--stage-path", stage_path])
-            try:
-                os.remove(stage_path)
-            except OSError:
-                pass
+            if stage_path:
+                command.extend(["--stage-path", stage_path])
+                try:
+                    os.remove(stage_path)
+                except OSError:
+                    pass
+        else:
+            # new-style example, setup viewer type and output path
+            if USD_AVAILABLE:
+                stage_path = os.path.join(
+                    os.path.dirname(__file__), f"outputs/{name}_{sanitize_identifier(device)}.usd"
+                )
+                command.extend(["--viewer", "usd", "--output-path", stage_path])
+            else:
+                stage_path = "None"
+                command.extend(["--viewer", "null"])
 
         command.extend(_build_command_line_options(options))
 
@@ -164,6 +176,10 @@ def add_example_test(
             result = subprocess.run(
                 command, capture_output=True, text=True, env=env_vars, timeout=test_timeout, check=False
             )
+
+        # print any error messages (e.g.: module not found)
+        if result.stderr != "":
+            print(result.stderr)
 
         # Check the return code (0 is standard for success)
         test.assertEqual(
@@ -186,16 +202,39 @@ cuda_test_devices = get_selected_cuda_test_devices(mode="basic")  # Don't test o
 test_devices = get_test_devices(mode="basic")
 
 
+class TestBasicExamples(unittest.TestCase):
+    pass
+
+
+add_example_test(TestBasicExamples, name="basic.example_basic_pendulum", devices=test_devices, use_viewer=True)
+
+add_example_test(
+    TestBasicExamples,
+    name="basic.example_basic_urdf",
+    devices=test_devices,
+    test_options_cpu={"num_envs": 16},
+    test_options_cuda={"num_envs": 64},
+    use_viewer=True,
+)
+
+add_example_test(TestBasicExamples, name="basic.example_basic_viewer", devices=test_devices, use_viewer=True)
+
+add_example_test(TestBasicExamples, name="basic.example_basic_joints", devices=test_devices, use_viewer=True)
+
+add_example_test(TestBasicExamples, name="basic.example_basic_shapes", devices=test_devices, use_viewer=True)
+
+
 class TestClothExamples(unittest.TestCase):
     pass
 
 
 add_example_test(
     TestClothExamples,
-    name="example_cloth_bending",
+    name="cloth.example_cloth_bending",
     devices=test_devices,
-    test_options={"usd_required": True, "stage_path": "None"},
+    test_options={"num_frames": 100},
     test_options_cpu={"num_frames": 100},
+    use_viewer=True,
 )
 add_example_test(
     TestClothExamples,
@@ -207,18 +246,20 @@ add_example_test(
 )
 add_example_test(
     TestClothExamples,
-    name="example_cloth_hanging",
+    name="cloth.example_cloth_hanging",
     devices=test_devices,
-    test_options={"stage_path": "None"},
+    test_options={},
     test_options_cpu={"width": 32, "height": 16, "num_frames": 10},
+    use_viewer=True,
 )
 add_example_test(
     TestClothExamples,
-    name="example_cloth_style3d",
+    name="cloth.example_cloth_style3d",
     devices=test_devices,
-    test_options={"usd_required": True, "stage_path": "None"},
-    test_options_cuda={"num_frames": 500},
+    test_options={},
+    test_options_cuda={"num_frames": 32},
     test_options_cpu={"num_frames": 2},
+    use_viewer=True,
 )
 
 
@@ -314,10 +355,12 @@ class TestOtherExamples(unittest.TestCase):
 
 add_example_test(
     TestOtherExamples,
-    name="example_mpm_granular",
+    name="mpm.example_mpm_granular",
     devices=cuda_test_devices,
-    test_options={"headless": True, "num_frames": 100},
+    test_options={"viewer": "null", "num_frames": 100},
+    use_viewer=True,
 )
+
 add_example_test(
     TestOtherExamples,
     name="example_rigid_force",
